remove custom build caching solution"
